requirements.txt
PyQt5
pysrt
pytube
requests
moviepy
googletrans
PyQt6-Frameless-Window

whisper_timestamped
